# Improve performance of haskell parser