--- a/attoparsec/src/Main.hs
+++ b/attoparsec/src/Main.hs
@@ -14,29 +14,10 @@ data BOX = FTYP B.ByteString Word32 [B.ByteString] | FREE | MOOV | SKIP | MDAT 
 
 ftypBox :: C.Parser BOX
 ftypBox = do
-  C.string "ftyp"
   major_brand         <- C.take 4
   major_brand_version <- Bin.anyWord32be
   compatible_brands   <- C.many' (C.take 4)
   return $ FTYP major_brand major_brand_version compatible_brands
-
-freeBox = do
-  C.string "free"
-  return  FREE
-mdat = do
-  C.string "mdat"
-  return  MDAT
-moov = do
-  C.string "moov"
-  return  MOOV
-skip = do
-  C.string "skip"
-  return SKIP
-mdra = C.string "mdra"
-dref = C.string "dref"
-cmov = C.string "cmov"
-rmra = C.string "rmra"
-
 
 mp4Box :: C.Parser LV
 mp4Box = do
@@ -44,7 +25,15 @@ mp4Box = do
   Right value  <- C.parseOnly anyBox <$> C.take (length - 4)
   return $ LV length value
 
-anyBox = C.choice [ftypBox, freeBox, moov, skip, mdat]
+anyBox = do
+  name <- C.take 4
+  case name of
+    "ftyp" -> ftypBox
+    "free" -> return FREE
+    "mdat" -> return MDAT
+    "moov" -> return MOOV
+    "skip" -> return SKIP
+    _      -> fail "Unknown box type"
 
 mp4Parser = C.many1 mp4Box