@Hrothen 2015-03-15
Turns out using a different parsing library gets the runtime down very close to nom. https://github.com/Codas/nom_benchmarks/tree/dev. Might be better to just swap that in entirely.