@Geal 2015-03-16
Thanks for your help! Multiple implementations are fine, it's good to compare everything. I guess at one point, this project will turn into "parser benchmarks" instead of just "nom benchmarks" :smile: